interaction with dashboard page

+++ routes/web.php
@@ -1,7 +1,11 @@
 <?php
 
 use App\Http\Controllers\ArticleController;
+// use App\Http\Controllers\UserController;
 use App\Http\Controllers\PagesController;
+use App\Http\Controllers\UserController;
+use App\Http\Controllers\Auth\RegisterController;
+use App\Http\Controllers\Auth\SessionsController;
 use Illuminate\Support\Facades\Route;
 
 /*
@@ -24,11 +28,22 @@ Route::controller(PagesController::class)->group(function () {
     Route::get('/about', 'about');
 });
 Route::controller(ArticleController::class)->group(function(){
-    Route::get('/articles','index')->name('articles.index');
-    Route::get('/articles/create','create')->name('articles.create');
-    Route::post('/articles','store')->name('articles.store');
-    Route::get('/articles/{article}','show')->name('articles.show');
-    Route::get('/articles/{article}/edit','edit')->name('articles.edit');    
-    Route::patch('/articles/{article}', 'update')->name('articles.update');
-    Route::delete('/articles/{article}', 'destroy')->name('articles.destroy');
+    Route::get('/articles','index')->name('articles.index')->middleware('auth');
+    Route::get('/articles/create','create')->name('articles.create')->middleware('auth');
+    Route::post('/articles/create','store')->name('articles.store')->middleware('auth');
+    Route::get('/articles/{article}','show')->name('articles.show')->middleware('auth');
+    Route::get('/articles/{article}/edit','edit')->name('articles.edit')->middleware('auth');    
+    Route::patch('/articles/{article}/edit', 'update')->name('articles.update')->middleware('auth');
+    Route::delete('/articles/{article}/delete', 'destroy')->name('articles.destroy')->middleware('auth');
 });     
+
+// routes d'authentification 
+
+Route::get('/register', [RegisterController::class, 'index'])->name('register')->middleware('guest');
+Route::post('/register', [RegisterController::class, 'create'])->middleware('guest');
+Route::get('/login', [SessionsController::class, 'index'])->name('login')->middleware('guest');
+Route::post('/login', [SessionsController::class, 'authenticate'])->middleware('guest');
+Route::get('/logout', [SessionsController::class, 'logout'])->name('logout')->middleware('auth');
+
+// // routes dirigeant vers le profil
+Route::get('/profile', [UserController::class, 'index'])->name('profile')->middleware('auth.basic');
